Creating a more rational methodology for locating partial files
The _partial.spark files will now be located in any parent folder, or Shared subfolder off of a parent folder.

This change is primarily to enable Views/AreaName/Shared/_file.spark to be located.

This is a superset of the previous current directory and root shared directory, so it is backward compatible in that sense.