build-distribution.cmd no longer requires wix or packageloadkey.h
WIX binaries are in the bin folder, so does not need to be installed
on the building workstation. An empty packageloadkey.h is also
created if the real one does not exist on disk. The resulting VSIP may
be installed on machines which have the VS SDK installed, which is good
for VSIP development and testing.