Binding secondary view rendering content/once collections directly
When RenderPartial was called from within a call to RenderPartial,
there could be some output loss if they both attempt to write to a
named content section that didn't exist before. This change binds
the partial views to the content and once dictionaries directly so new
entries are mutually available immediately.