Refactoring SourceBuilder as SourceWriter and adding app.config to tests
Using the code dom's method to get an instance by language has the
unfortunate side-effect of requiring a config section to ensure the
compiler used is the correct v3.5 version.

The SourceBuilder is also now obsolete and the SourceWriter is responsible
for escrow lines, indentation, and tracking source snippet mappings.