Adding a LanguageKit with more advanced implementation of lang filter
Installs a descriptor filter and a view folder wrapper which work
together to enable localized partial files. The descriptor creates an
entirely different path that starts with "language/xx-xx" and the
viewfolder wrapper projects an illusion of the entire views directory
into this location, adjusting file names as needed for the requested
language files to be the ones which are used.